+++ PIO.ServerLib/Modules/FunctionalModules/MoverModule.cs
@@ -53,7 +53,7 @@ namespace PIO.ServerLib.Modules
 			}
 
 			Log(LogLevels.Information, $"Creating task (WorkerID={WorkerID})");
-			task = Try(() => taskModule.InsertTask((int)TaskTypeIDs.MoveTo, WorkerID,TargetFactoryID, null,GetLastETA(WorkerID).AddSeconds(10))).OrThrow<PIOInternalErrorException>("Failed to create task");
+			task = Try(() => taskModule.InsertTask(TaskTypeIDs.MoveTo, WorkerID,TargetFactoryID, null,GetLastETA(WorkerID).AddSeconds(10))).OrThrow<PIOInternalErrorException>("Failed to create task");
 
 			OnTaskCreated(task);
